% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\docType{class}
\name{aggregation}
\alias{aggregation}
\title{Create a kriging   adapter class for the objects code{km} of package
of the package \code{DiceKriging}.}
\format{An \code{\link{R6Class}} generator object}
\usage{
aggregation
}
\description{
\code{krigingsm} is a kriging adapter. It makes the use of  the universal distirbution easier
}
\details{
It construct a kriging model from the package DiceKriging and it inherits form  the \code{sm}. 
The inheritance of the class  \code{sm} is required when one want to use the universal distribution
}
\section{Fields}{

\describe{
\item{\code{numsm}}{: number of sub models}

\item{\code{model_name}}{: "aggregation"}

\item{\code{aggweights}}{: aggregation weights}

\item{\code{fit}}{: fitness function used to tune the wieghts}
}}
\section{Methods}{


\describe{
  \item{\code{krigingsm$new(x=NULL,y=NULL,parameters = NULL, type= "UK")}}{Creates a new \code{krigingsm} object. }
  \item{\code{train()}}{train the surrogate model.}
 \item{\code{predict(newdata)}}{ it predicts the new data.}
}
}
\examples{
library(UP)
d            <- 2;
n            <- 16
X    	      <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
Xtest	      <- expand.grid(x1=seq(0,1,length=5), x2=seq(0,1,length=4))
Y            <- apply(X, 1, branin)
sm = krigingsm$new()
sm$setDOE(X,Y)
sm$train() 
sm$predict(Xtest)

   
}
\keyword{datasets}

