% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticfunctions.R
\name{branin}
\alias{Tfunc}
\alias{branin}
\title{analytical functions}
\usage{
branin(X)

Tfunc(X)
}
\arguments{
\item{X}{points coordinates}
}
\value{
\code{branin} - a real value.

\code{Tfunc} - a real value.
}
\description{
\code{branin} - A test function

\code{Tfunc} - A test function.
}

