% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fit.R
\docType{class}
\name{custom_fit}
\alias{custom_fit}
\title{Create a fitness class with \code{custom_fit}}
\format{An \code{\link{R6Class}} generator object}
\usage{
custom_fit
}
\description{
\code{custom_fit} is designed to estimate the quality of a surrogate
}
\details{
custom fit enable linear combination of several criteria
}
\section{Fields}{

\describe{
\item{\code{listComp}}{List of component}

\item{\code{dcoeff}}{Coefficient of each term fitness}
}}
\section{Methods}{


\describe{
  \item{\code{custom_fit$new(coeff, components)}}{Creates a new \code{custom_fit}. Used to construct an assesment of a quality of a surrogate model}
  \item{\code{get_val(surrogate_model)}}{It gives an evaluation of a surrogate model (object \code{sm}).}
}
}
\examples{
library(UP)
d              <- 2
X      	      <- expand.grid(x1=seq(0,1,length=5), x2=seq(0,1,length=4))
y        	    <- apply(X, 1, branin)
upsm     	    <- UPSM$new(sm= krigingsm$new(), UP=UPClass$new(X,y,Scale =TRUE))
listCrit       <- list()
listCrit[[1]]  <- MSE$new() 
listCrit[[2]]  <- Resampling_Error$new() 
listCrit[[3]]  <- penlrm$new(X, y)
PPS            <- custom_fit$new(c(4,2,1), listCrit)
valPPS         <- PPS$get_val(upsm)
print("PPS")
print(valPPS)

}
\keyword{datasets}

