% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.R
\docType{class}
\name{ga}
\alias{ga}
\title{Genetic Algorithm to select best metamodel \code{ga}}
\format{An \code{\link{R6Class}} generator object}
\usage{
ga
}
\description{
\code{ga} is designed to estimate the quality of a surrogate
}
\details{
Automatic selection for general surrogates
}
\section{Fields}{

\describe{
\item{\code{listComp}}{List of component}

\item{\code{dcoeff}}{Coefficient of each term fitness}
}}
\section{Methods}{


\describe{
  \item{\code{ga$new(x, y)}}{Creates a new \code{custom_fit}. Used to construct an assesment of a quality of a surrogate model}
  \item{\code{evaluate(newdta)}}{It gives an evaluation of a surrogate model (object \code{sm}).}
}
}
\examples{
library(UP)
d              <- 2
X              <- expand.grid(x1=seq(0,1,length=5), x2=seq(0,1,length=4))

}
\keyword{datasets}

