% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_crit.R
\name{inverse_crit}
\alias{inverse_crit}
\title{TMSE criteria}
\usage{
inverse_crit(x, model, target = 0.5, alpha = 0.1)
}
\arguments{
\item{x}{points of the design space in which the criterion will be computed}

\item{model}{the surrogate model}

\item{target}{the target value}

\item{alpha}{the value of the distance criteria default alpha = 0.001}
}
\value{
the value of the criterion
}
\description{
Compute the target mean squared error for a set of points of the design space
}
\examples{
#' library(UP)
d            <- 2
n            <- 16
X            <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
Xtest	      <- expand.grid(x1=seq(0,1,length=6), x2=seq(0,1,length=6))
Y            <- apply(X, 1, branin)
upsm         <- UPSM$new(sm= krigingsm$new(), UP=UPClass$new(X,Y,Scale =TRUE))
crit         <- inverse_crit(x= t(Xtest), model=upsm) 
print(max(crit))

}

