% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krigingsm.R
\docType{class}
\name{krigingsm}
\alias{krigingsm}
\title{Create a kriging   adapter class for the objects code{km} of package
 of the package \code{DiceKriging}.}
\format{An \code{\link{R6Class}} generator object}
\usage{
# kriging     <- krigingsm$new()
}
\description{
\code{krigingsm} is a kriging adapter. It makes the use of  the universal distirbution easier
}
\details{
It construct a kriging model from the package DiceKriging and it inherits form  the \code{sm}. 
The inheritance of the class  \code{sm} is required when one want to use the universal distribution
}
\section{Fields}{

\describe{
\item{\code{model_name}}{: Default "Kriging"}

\item{\code{type}}{: universal kriging "UK" or simple kriging "SK"}
}}
\section{Methods}{


\describe{
  \item{\code{krigingsm$new(x=NULL,y=NULL,parameters = NULL, type= "UK")}}{Creates a new \code{krigingsm} object. }
  \item{\code{train()}}{train the surrogate model.}
 \item{\code{predict(newdata)}}{ it predicts the new data.}
}
}
\examples{
library(UP)
d            <- 2;
X    	      <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
Xtest	      <- expand.grid(x1=seq(0,1,length=10), x2=seq(0,1,length=10))
Y            <- apply(X, 1, branin)
sm = krigingsm$new()
sm$setDOE(X,Y)
sm$train() 
predictions <- sm$predict(Xtest)


   
}
\keyword{datasets}

