% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.R
\docType{class}
\name{MSE}
\alias{MSE}
\title{Mean Square Error \code{MSE}}
\format{An \code{\link{R6Class}} generator object}
\usage{
MSE
}
\description{
\code{MSE} is designed to easily compute MSE on design points
}
\details{
Compute MSE of a surrogate model
}
\section{Methods}{


\describe{
  \item{\code{MSE$new()}}{Creates a new \code{MSE}}
  \item{\code{get_val(surrogate_model)}}{ It evaluates the MSE for a surrogate_model  (object \code{sm}).}
}
}
\examples{
library(UP)
d            <- 2
X    		    <- expand.grid(x1=seq(0,1,length=5), x2=seq(0,1,length=4))
y        	  <- apply(X, 1, branin)
upsm     	  <- UPSM$new(sm= krigingsm$new(), UP=UPClass$new(X,y,Scale =TRUE))
mse          <- MSE$new()          
valmse       <- mse$get_val(upsm)
print("MSE")
print(valmse)

     
}
\keyword{datasets}

