% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen.R
\docType{class}
\name{penlrm}
\alias{penlrm}
\title{Create a penalty class with \code{penlrm}}
\format{An \code{\link{R6Class}} generator object}
\usage{
penlrm
}
\description{
\code{penlrm} is designed to generate extra data and estimate its values piecewise linear. 
Its is often used as a penalty that penalizes high surrogate model oscillations.
}
\details{
Construct data points within some simplices of the design od ecperiments 
 and evaluate a \code{sm} object with regards to the descrepency between its prediction and the generated data
}
\section{Fields}{

\describe{
\item{\code{Xpen}}{the witness data generated to perform comparison}

\item{\code{Ypen}}{the piecewise linear estimation}
}}
\section{Methods}{


\describe{
  \item{\code{penlrm$new(x,y)}}{Creates a new \code{penlrm}. Often used to construct the penalty measure and generate its data.}
  \item{\code{get_val(surrogate_model)}}{ It evaluates the penalty for a surrogate_model (object \code{sm}).}
  \item{\code{get_err(surrogate_model)}}{ It evaluates the errors of a surrogate model (object \code{sm}).}
}
}
\examples{
library(UP)
X   <- matrix(1:40,20,2)  
Y   <- runif(20)
pen <- penlrm$new(X, Y)

x      	<- expand.grid(x1=seq(0,1,length=5), x2=seq(0,1,length=4))
y        <- apply(x, 1, branin)
upsm     <- UPSM$new(sm= krigingsm$new(), UP=UPClass$new(x,y,Scale =TRUE))

lrmval   <- pen$get_val(upsm)
print("val_lrm")
print(lrmval)

}
\keyword{datasets}

