% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFunc.R
\name{plotContour2D}
\alias{plotContour2D}
\title{contour plot of a function}
\usage{
plotContour2D(fun, partitions, nlevel = 40)
}
\arguments{
\item{fun}{a 2D function}

\item{partitions}{number of partirions per dimension}

\item{nlevel}{number of line levels}
}
\description{
contour plot of a function
}
\examples{
plotContour2D(branin,40,40)

}

