% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFunc.R
\name{plotUP1D}
\alias{plotUP1D}
\title{Post Processing and plot of the prediction and the UP variance 
of a 1 dimension functions.}
\usage{
plotUP1D(xverif, uppred, x = NULL, y = NULL)
}
\arguments{
\item{xverif}{verification points in 1 dimension.}

\item{uppred}{UP prediction of xverif.}

\item{x}{design points inputs.}

\item{y}{design points outputs.}
}
\value{
Plot the predictions.
}
\description{
Post Processing and plot of the prediction and the UP variance 
of a 1 dimension functions.
}
\examples{
library(UP)
x           <- as.matrix(c(-2.6,-0.2, 1.7,-1.4,1.2,3))
y           <- c(0.8, 0.5, 0.1, 0.3, 0, 0.4)
xverif      <- seq(-3, 3, length.out =300)
krig        <- krigingsm$new()
resampling  <- UPClass$new(x, y, Scale =TRUE) 
upsm        <- UPSM$new(sm= krig, UP= resampling) 
prediction  <- upsm$uppredict(xverif)
plotUP1D(xverif, prediction, x, y)

}

