% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_error.R
\docType{class}
\name{Resampling_Error}
\alias{Resampling_Error}
\title{Resampling Error \code{Resampling_Error}}
\format{An \code{\link{R6Class}} generator object}
\usage{
Resampling_Error
}
\description{
\code{Resampling_Error} is designed to easily compute resampling error on design points
}
\details{
Compute resampling error of a surrogate model
}
\section{Methods}{


\describe{
  \item{\code{Resampling_Error$new()}}{Creates a new \code{Resampling_Error}}
  \item{\code{get_val(surrogate_model)}}{ It evaluates the resampling error for a surrogate_model  (object \code{UPSM}).}
}
}
\examples{
library(UP)
d              <- 2
X      	      <- expand.grid(x1=seq(0,1,length=5), x2=seq(0,1,length=4))
y        	    <- apply(X, 1, branin)
upsm     	    <- UPSM$new(sm= krigingsm$new(), UP=UPClass$new(X,y,Scale =TRUE))
CV             <- Resampling_Error$new()          
PRESS_CV       <- CV$get_val(upsm)
print(" CV Resampling Error")
print(PRESS_CV)

     
}
\keyword{datasets}

