% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rwrapper.R
\name{kriging}
\alias{kriging}
\alias{masterprediction}
\alias{predict}
\alias{setDOE}
\alias{svm}
\alias{train}
\alias{uppredict}
\alias{upsm}
\title{Traditional R calls}
\usage{
kriging(x = NULL, y = NULL, parameters = NULL, type = "UK")

svm(x = NULL, y = NULL, parameters = NULL)

train(model)

predict(model, newdata)

upsm(sm = NULL, UP = NULL)

setDOE(model, x, y)

masterprediction(model, newdata)

uppredict(model, newdata)
}
\arguments{
\item{x}{input values}

\item{y}{output values}

\item{parameters}{kriging parameters}

\item{type}{kriging type}

\item{model}{the surrogate model}

\item{newdata}{data points where to predict}

\item{sm}{surrogate model}

\item{UP}{Resampling method defining the Universal Prediction}
}
\value{
kriging: kriging surrogate model.

svm: svm surrogate model

train: trained surrogate model

predict: surrogate model predictions

upsm: UP surrogate model

setDOE: a UP surrogate model with DOE

masterprediction: surrogate model master predictions

uppredict: surrogate model UP predictions
}
\description{
\code{kriging} - kriging surrogate model

\code{svm} - SVM surrogate models

\code{train} - train surrogate models

\code{predict} - prediction values of the  surrogate models

\code{upsm} - create a UP surrogate model

\code{setDOE} - set DOE for a UP surrogate model

\code{masterprediction} - master prediction values of a UP surrogate models

\code{uppredict} - UP prediction values of a UP surrogate models
}

