% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmsm.R
\docType{class}
\name{svmsm}
\alias{svmsm}
\title{Create a support vector machine (svm) adapter class for  the objects  code{svm} of  the package \code{e1071}.}
\format{An \code{\link{R6Class}} generator object}
\usage{
# svm     <- svmsm$new()
}
\description{
\code{svmsm} is a svm adapter. It makes the use of  the universal distirbution easier
}
\details{
It construct a svm model adapter  for  the objects  code{svm} of  the package \code{e1071}  and it inherits the class suurogate model \code{sm}. 
The inheritance of the class \code{sm} is required when one want to use the universal distribution
}
\section{Fields}{

\describe{
\item{\code{model_name}}{:"svm"}
}}
\section{Methods}{


\describe{
  \item{\code{svmsm$new(x=NULL,y=NULL,parameters = NULL )}}{Creates a new \code{svmsm} object. }
  \item{\code{train()}}{train the surrogate model.}
 \item{\code{predict(newdata)}}{ it predicts the new data.}
}
}
\examples{
library(UP)
d      <- 2
n      <- 16
X    	<- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
Xtest	<- expand.grid(x1=seq(0,1,length=5), x2=seq(0,1,length=4))
Y      <- apply(X, 1, branin)
sm     <- svmsm$new()

sm$setDOE(X,Y)
sm$train() 
sm$predict(Xtest)

   
}
\keyword{datasets}

