% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UP.R
\docType{package}
\name{UP}
\alias{UP}
\alias{UP-package}
\title{UP: Universal Prediction Distribution}
\description{
\code{UP} is a package that provides universal method for surrogate models: A package with extensible options and various 
UP-based algorithms for optimization, refienment and inversion.
}
\section{Introduction}{


Universal Prediction Distribution for surrogate models is an efficient too for adaptive designs. It allows the extension
of many algorithms defined for the Gaussian case for all types of surrogates
}

\section{UP Distribution}{


\code{UP}  of a surogate model is given by the prediction of its submodels defined by a resampling technique.
}

\section{Methods}{


The package also contains several algorihms that uses the UP distribution such as:
 \code{\link{upego}},   \code{\link{upsmart}},  \code{\link{upinverse}} and \code{\link{ga}} 
model selection.
}

\section{Applications}{


\code{UP} applications are diverse. We can cite  mainely: sequential sampling, optimization and model selection.
\code{UP}is always useful when one uses surrogate modeling paradigm.
}
\examples{

library(UP)
d           <- 2;
n           <- 16
testdata    <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
design.fact <- expand.grid(x1=seq(0,1,length=5), x2=seq(0,1,length=4))
y           <- apply(design.fact, 1, branin)
aUP         <- UPClass$new(design.fact,y,Scale =TRUE, resampling_type = "KFCV", kfold =5)
kriging     <- krigingsm$new()
upkrig      <- UPSM$new(sm= kriging, UP =aUP)
predicted.values.model1 <- upkrig$uppredict(testdata)
max(predicted.values.model1$upsd)
upkrig2     <- UPSM$new(sm= krigingsm$new(), UP=UPClass$new(design.fact,y,Scale =TRUE))
uppred2     <- upkrig2$uppredict(testdata)
max(uppred2$upsd)
upsvm       <- UPSM$new(sm= svmsm$new(), UP=UPClass$new(design.fact,y,Scale =TRUE))
uppred3     <- upsvm$uppredict(testdata)
uppred3

}
\seealso{
\code{\link{UP}}
}

