% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPClass.R
\docType{class}
\name{UPClass}
\alias{UPClass}
\title{A UPClass is a class the contains a resampling technique and computes the weights of the UP distirbution}
\format{An \code{\link{R6Class}} generator object}
\usage{
# up     <- UPClass$new()
}
\description{
\code{UPClass} is a class for the resampling technique and its use in the UP disitrbution paradigm.
}
\details{
It scale the data, initilize a resmapling technique and compute the weigths of the UP distirbution
}
\section{Fields}{

\describe{
\item{\code{Resampling:}}{a resampling technqiue (Default leva-one-out LOO)}

\item{\code{x}}{: design points}

\item{\code{y}}{: response}

\item{\code{Xscaled}}{: Scaled design points}

\item{\code{Yscaled}}{: Sacelde response}

\item{\code{d}}{: dimension}

\item{\code{n}}{: number of design points}

\item{\code{rho:}}{weight parameters}

\item{\code{variogramme:}}{weight paramter (default "gauss")}
}}
\section{Methods}{
 
\describe{
  \item{\code{UPClass$new(x = NA, y = NA, Scale =TRUE,
				resampling_type = "LOO", ListIndices = NULL, 
				kfold = 10, division_method ="random") }}{Creates a new \code{UPClass} object. }
  \item{\code{scaleinput(newData)}}{ Scale new data.}
  \item{\code{unscaleoutput(values )}}{unscale scaled output}
 \item{\code{computeweights(scnewdata)}}{ compute the up weights for a new set of data.}
}
}
\examples{
library(UP)
d          	<- 2;
n           	<- 16
X    		    <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
y          	<- apply(X, 1, branin)
aUP        	<- UPClass$new(X,y,Scale =TRUE, resampling_type = "KFCV", kfold =5)

   
}
\keyword{datasets}

