% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upego.R
\name{upego}
\alias{upego}
\title{Universal Prediction distribution based  Efficient Global Optimization UP-EGO}
\usage{
upego(model, fun, nsteps, lower, upper, seed = 1, parinit = NULL,
  control = NULL, EEIcontrol = NULL)
}
\arguments{
\item{model}{the surrogate model}

\item{fun}{the real function}

\item{nsteps}{the number of points to be generated}

\item{lower}{the lower bound of the design space}

\item{upper}{the upper bound of the design space}

\item{seed}{the random seed (default = 1)}

\item{parinit}{inital points to be used in the optimization (default NULL)}

\item{control}{the optimization control parameters (default NULL)}

\item{EEIcontrol}{the optimization criterion parameters (default NULL)}
}
\value{
list of generated points and their values and the last updated surrogate model
}
\description{
Global optimization using the  Universal Prediction distribution
}
\examples{
#' library(UP)
d            <- 2;
n            <- 16
X            <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
Y            <- apply(X, 1, branin)

upsm     	<- UPSM$new(sm= krigingsm$new(), UP=UPClass$new(X,Y,Scale =TRUE) ) 
upego_res <-  upego( upsm,fun = branin, nsteps = 1, lower= c(0,0),upper = c(1,1) )

print( min(upego_res$last$get_DOE()$y ) ) 

}

