% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPEI.R
\name{UPEI}
\alias{UPEI}
\title{Universal Prediction Expected Improvement UP-EI criteria}
\usage{
UPEI(x, model, plugin = NULL, alpha = 0.01, up_method = "Empirical",
  envir = NULL)
}
\arguments{
\item{x}{points of the design space in which the criterion will be computed}

\item{model}{the surrogate model}

\item{plugin}{the current minimum value}

\item{alpha}{a parameter for exploration term}

\item{up_method}{type of computation, default  "Empirical".  Other options: "UP_reg" regularized emprirical
or "assume_gauss"  regularize UP as Gaussian ditribution.}

\item{envir}{envirement variable}
}
\value{
the value of the UP expected improvement
}
\description{
Compute the universal preidction expected improvement
}
\examples{
#' library(UP)
d            <- 2
n            <- 16
X            <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
Xtest        <- expand.grid(x1=seq(0,1,length=6), x2=seq(0,1,length=6))
Y            <- apply(X, 1, branin)
upsm         <- UPSM$new(sm= krigingsm$new(), UP=UPClass$new(X,Y,Scale =TRUE))
crit         <- UPEI(x= t(Xtest), model=upsm,plugin=min(Y)) 
print(max(crit))

}

