% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upinverse.R
\name{upinverse}
\alias{upinverse}
\title{upinverse}
\usage{
upinverse(model, fun, target, nsteps, lower, upper, seed = 1,
  parinit = NULL, control = NULL, RefControl = NULL)
}
\arguments{
\item{model}{the surrogate model}

\item{fun}{the real function}

\item{target}{a target value}

\item{nsteps}{the number of points to be generated}

\item{lower}{the lower bound of the design space}

\item{upper}{the upper bound of the design space}

\item{seed}{the random seed (default = 1)}

\item{parinit}{inital points to be used in the optimization (default NULL)}

\item{control}{the optimization control parameters (default NULL)}

\item{RefControl}{the refienement criterion parameters (default NULL)}
}
\value{
list of generated points and their values and the last updated surrogate model
}
\description{
Generate  points to determine a line level
}
\examples{
library(lhs)
d            <- 2
n            <- 25
X            <- randomLHS(n,d)
Y            <- apply(X, 1, Tfunc)
upsm         <- UPSM$new(sm= krigingsm$new(), UP=UPClass$new(X,Y,Scale =TRUE)) 
upinv_res    <-  upinverse(upsm, fun = Tfunc, target = 0.8,
                   nsteps = 15, lower= c(0,0),upper = c(1,1))

}

