% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsmart.R
\name{upsmart}
\alias{upsmart}
\title{UP-SMART}
\usage{
upsmart(model, fun, nsteps, lower, upper, seed = 1, parinit = NULL,
  control = NULL, RefControl = NULL)
}
\arguments{
\item{model}{the surrogate model}

\item{fun}{the real function}

\item{nsteps}{the number of points to be generated}

\item{lower}{the lower bound of the design space}

\item{upper}{the upper bound of the design space}

\item{seed}{the random seed (default = 1)}

\item{parinit}{inital points to be used in the optimization (default NULL)}

\item{control}{the optimization control parameters (default NULL)}

\item{RefControl}{the refienement criterion parameters (default NULL)}
}
\value{
list of generated points and their values and the last updated surrogate model
}
\description{
Enhance the overall prediction capabilities of a surrogate model by the Universal Prediction 
distribution based Surrogate Modeling Adapative Refinement Strategy UP-SMART
}
\examples{
#' library(UP)
d            <- 2;
n            <- 16
X            <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
Xtest        <- expand.grid(x1=seq(0,1,length=5), x2=seq(0,1,length=4))
Y            <- apply(X, 1, branin)
sm           <- krigingsm$new()
sm$setDOE(X,Y)
sm$train() 
upsmart_res  <- upsmart(sm,fun = branin,nsteps = 5, lower= c(0,0),upper = c(1,1))

print(upsmart_res$last$get_DOE())

}

