% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{upsurface}
\alias{upsurface}
\title{Adapter for UP surface}
\usage{
upsurface(x, y, rs, upcontrol = NULL)
}
\arguments{
\item{x}{design input}

\item{y}{design output}

\item{rs}{the surrogate model}

\item{upcontrol}{list of resampling technique default NULL}
}
\value{
A surrogate model ready to be used in a UP framework
}
\description{
A surrogate model ready to be used in a UP framework
}
\examples{
library(UP)
d            <- 2
X            <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
testdata     <- expand.grid(x1=s <- seq(0,1, length=10), x2=s)
Y            <- apply(X, 1, branin)
upsm         <- upsurface(x=X, y=Y, rs= krigingsm$new())
predictions  <- upsm$uppredict(testdata)

}
